Created an external class for simple view animations
